refactor(folder): 重构文件夹模块权限控制逻辑
- 在 FolderController 中添加用户权限检查，仅允许 ADMIN 角色执行增删改操作
- 优化 FolderService 中的 tree 方法，简化根目录创建逻辑
- 新增 addFolder 和 updateFolder 方法，用于处理文件夹创建和更新
- 删除了未使用的 UserService 依赖

+++ src/main/java/com/tarzan/maxkb4j/module/folder/controler/FolderController.java
@@ -3,10 +3,10 @@ package com.tarzan.maxkb4j.module.folder.controler;
 import cn.dev33.satoken.stp.StpUtil;
 import com.tarzan.maxkb4j.constant.AppConst;
 import com.tarzan.maxkb4j.core.api.R;
-import com.tarzan.maxkb4j.module.folder.entity.FolderEntity;
 import com.tarzan.maxkb4j.module.folder.service.FolderService;
 import com.tarzan.maxkb4j.module.folder.vo.FolderVO;
-import com.tarzan.maxkb4j.util.BeanUtil;
+import com.tarzan.maxkb4j.module.system.user.domain.entity.UserEntity;
+import com.tarzan.maxkb4j.module.system.user.service.UserService;
 import lombok.AllArgsConstructor;
 import org.springframework.web.bind.annotation.*;
 
@@ -18,27 +18,40 @@ import java.util.List;
 public class FolderController {
 
     private final FolderService folderService;
+    private final UserService userService;
 
 
     @PostMapping("/{source}/folder")
     public R<Boolean> addFolder(@PathVariable String source, @RequestBody FolderVO folder) {
-        FolderEntity folderEntity = BeanUtil.copy(folder, FolderEntity.class);
-        folderEntity.setUserId(StpUtil.getLoginIdAsString());
-        folderEntity.setSource(source);
-        return R.data(folderService.save(folderEntity));
+        UserEntity user = userService.getById(StpUtil.getLoginIdAsString());
+        if (user != null) {
+            if (user.getRole().contains("ADMIN")) {
+                return R.data(folderService.addFolder(source,folder));
+            }
+        }
+        return R.fail("无权限");
     }
 
     @PutMapping("/{source}/folder/{id}")
     public R<Boolean> updateFolder(@PathVariable String source, @PathVariable String id, @RequestBody FolderVO folder) {
-        FolderEntity folderEntity = BeanUtil.copy(folder, FolderEntity.class);
-        folderEntity.setId(id);
-        folderEntity.setSource(source);
-        return R.data(folderService.updateById(folderEntity));
+        UserEntity user = userService.getById(StpUtil.getLoginIdAsString());
+        if (user != null) {
+            if (user.getRole().contains("ADMIN")) {
+                return R.data(folderService.updateFolder(source,id,folder));
+            }
+        }
+        return R.fail("无权限");
     }
 
     @DeleteMapping("/{source}/folder/{id}")
     public R<Boolean> deleteFolder(@PathVariable String source, @PathVariable String id) {
-        return R.data(folderService.deleteFolder(source, id));
+        UserEntity user = userService.getById(StpUtil.getLoginIdAsString());
+        if (user != null) {
+            if (user.getRole().contains("ADMIN")) {
+                return R.data(folderService.deleteFolder(source, id));
+            }
+        }
+        return R.fail("无权限");
     }
 
     @GetMapping("/{source}/folder")
